blocklistd-helper: make backend detection more reliable
Using the existence of a configuration file to determine which packet
filter subsystem is in use is a losing game and liable to fail on
systems which can support multiple different packet filters.

Note also that /etc/netscript/ipfilter.conf on GNU/Linux is not anything
to do with IPFilter.  It's just a shell script fragment for netscript(8).
IPFilter hasn't been available for Linux kernels for quite some time --
it probably has not worked since 2.6, some decades ago, though Phil
Dibowitz's IP Filter FAQ hasn't been updated since 2009.

This change will use the normal BSD /etc/rc.d/ scripts, with the status
command, to detect which firewall is enabled; and only if there is no
such script then it will look for the likely configuration file.

GNU/Linux systems will have to do with the existing testing for the
presence of the /sbin/iptables executable.

Using the rc.d script to detect IPFilter necessarily means using
"ipfilter" as the internal tag for it to keep things simple enough.
This is good because that's the name of the packet filter
subsystem ("ipf" is just the name of the main command interface and the
root of its default configuration file on most BSD systems).

Finally this change also includes some updates to the comments about
IPFilter usage within.